Added Activity Tab Filter!
Now, I have added the function of filterActivityRowsByHref. With this function you can change your activity tab to only show people you want to see in your activity. It is turned off by default, all you have to do to turn it on is to remove the hashtag besides the function. You can also choose to show reviews regardless of the user.